+++ components/webui/server/package.json
@@ -6,13 +6,13 @@
   "scripts": {
     "build": "tsc",
     "build:watch": "npm run build -- -w",
-    "dev:start": "fastify start --ignore-watch=.ts$ -w -l info -P dist/server/src/fastify-v2/app.js",
+    "dev:start": "fastify start --ignore-watch=.ts$ -w -l info -P dist/server/src/app.js",
     "dev": "npm run build && concurrently -k -p \"[{name}]\" -n \"TypeScript,App\" -c \"yellow.bold,cyan.bold\" \"npm:build:watch\" \"npm:dev:start\"",
-    "start": "npm run build && fastify start -l info dist/server/src/fastify-v2/app.js",
+    "start": "npm run build && fastify start -l info dist/server/src/app.js",
     "standalone": "npm run build && node --env-file=.env dist/server/src/main.js",
     "lint:check": "eslint . --max-warnings 0",
     "lint:fix": "npm run lint:check -- --fix",
-    "test": "NODE_ENV=test tap --include='**/*test.ts'"
+    "test": "tap --include='**/*test.ts'"
   },
   "author": "YScope Inc. <dev@yscope.com>",
   "license": "Apache-2.0",
