+++ components/webui/server/src/plugins/app/QueryJobDbManager/typings.ts
