+++ components/webui/server/src/plugins/app/socket/MongoSocketIoServer/MongoWatcherCollection.ts
@@ -4,7 +4,7 @@ import type {
     Db,
 } from "mongodb";
 
-import {QueryId} from "../../../../../../../common/index.js";
+import {QueryId} from "../../../../../../common/index.js";
 import {
     CLIENT_UPDATE_TIMEOUT_MILLIS,
     MongoCustomSocket,
