+++ components/webui/server/src/plugins/app/socket/MongoSocketIoServer/typings.ts
@@ -11,7 +11,7 @@ import {
     InterServerEvents,
     ServerToClientEvents,
     SocketData,
-} from "../../../../../../../common/index.js";
+} from "../../../../../../common/index.js";
 
 
 /**
