+++ components/webui/server/src/routes/api/search/utils.ts
@@ -1,6 +1,6 @@
 import type {Db} from "mongodb";
 
-import {SEARCH_SIGNAL} from "../../../../../../common/index.js";
+import {SEARCH_SIGNAL} from "../../../../../common/index.js";
 import {
     CreateMongoIndexesProps,
     SEARCH_MAX_NUM_RESULTS,
