+++ components/webui/server/src/routes/static.ts
@@ -1,5 +1,4 @@
 import path from "node:path";
-import process from "node:process";
 import {fileURLToPath} from "node:url";
 
 import {fastifyStatic} from "@fastify/static";
@@ -38,26 +37,22 @@ const routes: FastifyPluginAsync = async (fastify) => {
         decorateReply: false,
     });
 
-    if ("production" === process.env.NODE_ENV) {
-        // In the development environment, we expect the client to use a separate webserver that
-        // supports live reloading.
-        let clientDir = settings.ClientDir;
-        if (false === path.isAbsolute(clientDir)) {
-            clientDir = path.resolve(rootDirname, settings.ClientDir);
-        }
-
-        await fastify.register(fastifyStatic, {
-            prefix: "/",
-            root: clientDir,
-            decorateReply: true,
-            wildcard: false,
-        });
-
-        // Serve index.html for all unmatched routes in the React Single Page Application (SPA).
-        fastify.get("/*", (_, reply) => {
-            reply.sendFile("index.html");
-        });
+    let clientDir = settings.ClientDir;
+    if (false === path.isAbsolute(clientDir)) {
+        clientDir = path.resolve(rootDirname, settings.ClientDir);
     }
+
+    await fastify.register(fastifyStatic, {
+        prefix: "/",
+        root: clientDir,
+        decorateReply: true,
+        wildcard: false,
+    });
+
+    // Serve index.html for all unmatched routes in the React Single Page Application (SPA).
+    fastify.get("/*", (_, reply) => {
+        reply.sendFile("index.html");
+    });
 };
 
 export default routes;
