+++ components/webui/server/src/schemas/error.ts
