+++ components/webui/server/src/schemas/search.ts
