+++ components/webui/server/src/schemas/stream-files.ts
@@ -1,6 +1,6 @@
 import {Type} from "@sinclair/typebox";
 
-import {QUERY_JOB_TYPE} from "../../typings/query.js";
+import {QUERY_JOB_TYPE} from "../typings/query.js";
 import {StringSchema} from "./common.js";
